Add Habits and Goals features (full vertical slices)
Habits: six habit types (Yes/No, Counter, Timer, Duration, Checklist,
Collection) sharing one numeric progressValue/targetValue storage model,
atomic per-period upserts, streak/completion/heatmap stats, and list/
form/detail UI wired into the router and dashboard.

Goals: milestone-based or numeric-fallback progress tracking, with an
optional many-to-many link to Habits via a join table, plus list/form/
detail UI wired into the router and dashboard.

Also fixes a latent num.clamp() type bug (clamp returns num, not
double/int) across both new and previously-committed files (loan.dart,
card_emi.dart, recurrence_frequency.dart, labeled_progress_bar.dart,
rest_timer_banner.dart) that would have failed to compile.

Adds unit tests for HabitStats/GoalStats pure logic, CreateHabit/
CreateGoal/AddMilestone use-case validation, and HabitsDao/GoalsDao
atomic behavior against an in-memory database.

+++ lifeos/lib/core/widgets/labeled_progress_bar.dart
@@ -36,7 +36,7 @@ class LabeledProgressBar extends StatelessWidget {
         ClipRRect(
           borderRadius: BorderRadius.circular(8),
           child: LinearProgressIndicator(
-            value: progress.clamp(0, 1),
+            value: progress.clamp(0, 1).toDouble(),
             minHeight: 8,
             backgroundColor: trackColor,
             valueColor: color != null ? AlwaysStoppedAnimation(color) : null,
